Need tests!

+++ server/utils/fileAndFolder.ts
@@ -7,6 +7,7 @@ export const mkdirIfNotExist = (path: string): void => {
 };
 
 export const deletePath = (path: string): void => {
+	if (!path || !fs.existsSync(path)) return;
 	fs.rm(path, { recursive: true, force: true }, err => {
 		if (err) { throw err; }
 	});
